Iniciando camada de servico e implementando userService

+++ ControleAcesso.Domain/Interfaces/IFunctionalityProfileRepository.cs
@@ -5,8 +5,8 @@ namespace ControleAcesso.Domain.Interfaces
     public interface IFunctionalityProfileRepository : IGenericRepository<FunctionalityProfile>
     {
         Task<IEnumerable<FunctionalityProfile>> GetAllAsync();
-        Task<FunctionalityProfile> GetByIdAsync(int id);
-        Task<IEnumerable<FunctionalityProfile>> GetByProfileIdAsync(int profileId);
-        Task<IEnumerable<FunctionalityProfile>> GetByFunctionalityIdAsync(int functionalityId);
+        Task<FunctionalityProfile> GetByIdAsync(long id);
+        Task<IEnumerable<FunctionalityProfile>> GetByProfileIdAsync(long profileId);
+        Task<IEnumerable<FunctionalityProfile>> GetByFunctionalityIdAsync(long functionalityId);
     }
 }
