+++ ControleAcesso.Domain/Interfaces/IFunctionalityRepository.cs
@@ -5,6 +5,6 @@ namespace ControleAcesso.Domain.Interfaces
     public interface IFunctionalityRepository : IGenericRepository<Functionality>
     {
         Task<IEnumerable<Functionality>> GetAllAsync();
-        Task<Functionality> GetByIdAsync(int id);
+        Task<Functionality> GetByIdAsync(long id);
     }
 }
