+++ ControleAcesso.Domain/Interfaces/IMenuOptionRepository.cs
@@ -5,7 +5,7 @@ namespace ControleAcesso.Domain.Interfaces
     public interface IMenuOptionRepository : IGenericRepository<MenuOption>
     {
         Task<IEnumerable<MenuOption>> GetAllAsync();
-        Task<MenuOption> GetByIdAsync(int id);
-        Task<IEnumerable<MenuOption>> GetByFunctionalityIdAsync(int functionalityId);
+        Task<MenuOption> GetByIdAsync(long id);
+        Task<IEnumerable<MenuOption>> GetByFunctionalityIdAsync(long functionalityId);
     }
 }
