+++ ControleAcesso.Domain/Interfaces/IMethodRepository.cs
@@ -5,7 +5,7 @@ namespace ControleAcesso.Domain.Interfaces
     public interface IMethodRepository : IGenericRepository<Methods>
     {
         Task<IEnumerable<Methods>> GetAllAsync();
-        Task<Methods> GetByIdAsync(int id);
-        Task<IEnumerable<Methods>> GetByFunctionalityIdAsync(int functionalityId);
+        Task<Methods> GetByIdAsync(long id);
+        Task<IEnumerable<Methods>> GetByFunctionalityIdAsync(long functionalityId);
     }
 }
