+++ ControleAcesso.Domain/Interfaces/IUserProfileRepository.cs
@@ -5,8 +5,8 @@ namespace ControleAcesso.Domain.Interfaces
     public interface IUserProfileRepository : IGenericRepository<UserProfile>
     {
         Task<IEnumerable<UserProfile>> GetAllAsync();
-        Task<UserProfile> GetByIdAsync(int id);
-        Task<IEnumerable<UserProfile>> GetByUserIdAsync(int userId);
-        Task<IEnumerable<UserProfile>> GetByProfileIdAsync(int profileId);
+        Task<UserProfile> GetByIdAsync(long id);
+        Task<IEnumerable<UserProfile>> GetByUserIdAsync(long userId);
+        Task<IEnumerable<UserProfile>> GetByProfileIdAsync(long profileId);
     }
 }
