+++ ControleAcesso.Domain/Interfaces/IUserRepository.cs
@@ -5,7 +5,7 @@ namespace ControleAcesso.Domain.Interfaces
     public interface IUserRepository : IGenericRepository<User>
     {
         Task<IEnumerable<User>> GetAllAsync();
-        Task<User> GetByIdAsync(int id);
+        Task<User> GetByIdAsync(long id);
         Task<User> GetByUserNameAsync(string userName);
         Task<User> GetByEmailAsync(string email);
     }
